backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
python-multipart
ollama
lancedb
docling
soundfile
numpy
pydantic==2.10.6
pydantic[email]
pydantic-settings
python-dotenv
aiofiles
sqlalchemy
asyncpg
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
celery
redis
onnxruntime
onnx
email-validator
kokoro-onnx
sherpa-onnx
requests
deepface
mediapipe
ultralytics
opencv-python-headless
tf-keras
